# openapi.yaml
openapi: 3.0.0
info:
  title: OpenMind API
  version: 1.0.0
paths:
  /project/{project_slug}/note:
    post:
      summary: Create a new note
      operationId: createNote
      parameters:
        - name: project_slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
      responses:
        '201':
          description: Note created successfully
          headers:
            Location:
              description: URL of the created note
              schema:
                type: string
                example: /note/123e4567-e89b-12d3-a456-426614174000
        '500':
          description: Internal server error
  /project/{project_id}/notes:
    get:
      summary: Fetch all notes by project ID
      operationId: fetchNotesByProject
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
        '404':
          description: Project not found
  /notes/{note_id}:
    get:
      summary: Fetch a note by its ID
      operationId: fetchNoteById
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Note not found
    delete:
      summary: Scratch a note by its ID
      operationId: scratchNote
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Note scratched successfully
        '404':
          description: Note not found
  /project/create:
    post:
      summary: Create a new project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created successfully
          headers:
            Location:
              description: URL of the created project
              schema:
                type: string
                example: /project/new-project
        '409':
          description: Project already exists
        '500':
          description: Internal server error
  /project/{project_slug}/thought:
    post:
      summary: Create a new thought
      operationId: createThought
      parameters:
        - name: project_slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateThoughtRequest'
      responses:
        '201':
          description: Thought created successfully
          headers:
            Location:
              description: URL of the created thought
              schema:
                type: string
                example: /thought/123e4567-e89b-12d3-a456-426614174000
        '500':
          description: Internal server error
components:
  schemas:
    Note:
      type: object
      properties:
        id:
          type: integer
        imported_at:
          type: string
          format: date-time
        scribe_id:
          type: integer
        project_id:
          type: integer
        content:
          type: string
    CreateNoteRequest:
      type: object
      properties:
        imported_at:
          type: string
          format: date-time
        scribe_id:
          type: string
          format: uuid
        content:
          type: string
      required:
        - imported_at
        - scribe_id
        - content
    CreateProjectRequest:
      type: object
      properties:
        project_name:
          type: string
        universe_id:
          type: string
          format: uuid
      required:
        - project_name
        - universe_id
    CreateThoughtRequest:
      type: object
      properties:
        imported_at:
          type: string
          format: date-time
        scribe_id:
          type: string
          format: uuid
        content:
          type: string
      required:
        - imported_at
        - scribe_id
        - content